requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-telegram-bot>=20.0
notion-client>=2.0.0
python-dotenv>=1.0.0
